requirements.txt
pyspark==3.4.1
pandas>=1.5.0
psycopg2-binary>=2.9.0
minio>=7.1.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
jupyter>=1.0.0
matplotlib>=3.6.0
seaborn>=0.12.0